# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PDD Petty Cash API
  description: |
    REST API for the PDD Petty Cash (Travel Expense Management) app.
    See CLAUDE.md §9 for conventions: camelCase JSON, RFC 7807 errors, cursor pagination,
    Idempotency-Key required on POST /expenses, /transfers, /allocations.
  version: 0.1.0
servers:
  - url: https://api.pettycash.pdd.gov.ae/api/v1
    description: Production (Moro Hub) — placeholder
  - url: http://localhost:8080/api/v1
    description: Local Spring Boot
...
security:
  - bearerAuth: []
paths:
...
  /notifications:
    get:
      tags: [notifications]
      parameters:
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 30 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Notification' }
  /notifications/poll:
    get:
      tags: [notifications]
      summary: Long-poll for new notifications
      description: |
        Returns immediately if any notifications exist with
        `createdAt > since` for the current user. Otherwise the request hangs
        as an async `DeferredResult` for up to `timeoutSeconds` (default 25,
        max 30) and resolves when a new notification is published, or with an
        empty `items` array on timeout.

        Send the response's `serverNow` back as the next call's `since` to
        avoid client clock skew bouncing the window.

        Single-instance fan-out only — horizontal scale will require Redis
        Pub/Sub or Postgres LISTEN/NOTIFY (out of scope for v1).
      parameters:
        - in: query
          name: since
          description: ISO-8601 timestamp; only notifications strictly after this are returned. Defaults to now − 30s.
          schema: { type: string, format: date-time }
        - in: query
          name: timeoutSeconds
          schema: { type: integer, default: 25, minimum: 1, maximum: 30 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, serverNow]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Notification' }
                  serverNow:
                    type: string
                    format: date-time
  /notifications/{notificationId}/read:
    patch:
      tags: [notifications]
      parameters:
        - in: path
          name: notificationId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Notification' }
...
  /reports/trip/{tripId}:
    get:
      tags: [reports]
      summary: Generate a report and return a 5-minute presigned download URL
      description: |
        Renders the report server-side, uploads bytes to MinIO under
        `reports/{tripId}/{type}-{timestamp}.{ext}`, writes an `AuditLog`
        row with the file's SHA-256, and returns a presigned GET URL.

        Format gating per CLAUDE.md §10:
        - `USER` — XLSX or PDF
        - `TRIP` — XLSX only
        - `FINANCE` — PDF only; watermarked "DRAFT — unsigned" pending ADR-003
        - `DG` — PDF only

        Permissions (§7): USER → own (LEADER may fetch any team member),
        TRIP → LEADER (own trip) or ADMIN, FINANCE → ADMIN only,
        DG → SUPER_ADMIN only. Enforced at both controller and service.
      parameters:
        - { $ref: '#/components/parameters/TripIdPath' }
        - in: query
          name: type
          required: true
          schema: { type: string, enum: [USER, TRIP, FINANCE, DG] }
        - in: query
          name: format
          required: true
          schema: { type: string, enum: [XLSX, PDF] }
        - in: query
          name: userId
          description: Optional. Restricts USER/DG reports to a single user.
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [reportId, url, expiresAt, sha256]
                properties:
                  reportId: { type: string, format: uuid }
                  url:
                    type: string
                    description: Presigned GET URL valid for 5 minutes.
                  expiresAt: { type: string, format: date-time }
                  sha256: { type: string, description: SHA-256 hex of the bytes uploaded to MinIO. }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /reports/{reportId}/sign:
    post:
      tags: [reports]
      summary: Digitally sign a generated report (deferred — see ADR-003)
      description: |
        Signing is still deferred pending the PDD HSM key-custody decision.
        Endpoint returns `501 Not Implemented` with stable code
        `SIGNING_DEFERRED`. The contract is fixed so mobile can be wired
        against the final shape today.
      parameters:
        - in: path
          name: reportId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '501':
          description: Signing deferred
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              examples:
                signingDeferred:
                  summary: SIGNING_DEFERRED (501)
                  value:
                    code: SIGNING_DEFERRED
                    detail: "Signing key custody pending — see ADR-003."
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TripIdPath:
      in: path
      name: tripId
      required: true
      schema: { type: string }
...
  responses:
    Problem:
      description: RFC 7807 problem details with a stable `code` field per CLAUDE.md §9.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            idempotencyKeyRequired:
              summary: IDEMPOTENCY_KEY_REQUIRED (400)
              value:
                type: https://api.pettycash.pdd.gov.ae/problems/idempotency_key_required
                title: Bad Request
                status: 400
                detail: "Header 'Idempotency-Key' is required on this endpoint"
                instance: /api/v1/trips/abc/expenses
                code: IDEMPOTENCY_KEY_REQUIRED
            idempotencyKeyConflict:
              summary: IDEMPOTENCY_KEY_CONFLICT (409)
              value:
                type: https://api.pettycash.pdd.gov.ae/problems/idempotency_key_conflict
                title: Conflict
                status: 409
                detail: "Idempotency-Key reused with a different request body"
                instance: /api/v1/trips/abc/expenses
                code: IDEMPOTENCY_KEY_CONFLICT
            reportDeferred:
              summary: REPORT_DEFERRED (501)
              value:
                type: https://api.pettycash.pdd.gov.ae/problems/report_deferred
                title: Not Implemented
                status: 501
                detail: "Server-side report generation pending — see CLAUDE.md §10."
                instance: /api/v1/reports/trip/abc
                code: REPORT_DEFERRED
  schemas:
...
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
...
    NotificationType:
      type: string
      enum: [ALLOCATION_RECEIVED, TRANSFER_RECEIVED, TRANSFER_ACCEPTED, TRIP_ASSIGNED, TRIP_CLOSED, EXPENSE_QUERY]
    NotificationState:
      type: string
      enum: [UNREAD, READ, ACTED]
    Notification:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        type: { $ref: '#/components/schemas/NotificationType' }
        payload:
          type: object
          additionalProperties: true
        actionable: { type: boolean }
        state: { $ref: '#/components/schemas/NotificationState' }
        createdAt: { type: string, format: date-time }